backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.9.0
faker>=25.0.0
strawberry-graphql[fastapi]>=0.243.0
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
anyio[trio]>=4.4.0
